pkg/location: move command logic out of cmd/place to make it reusable
